requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.1
astroid==3.3.11
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
boto3==1.39.4
botocore==1.39.4
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.9.2
cron-descriptor==1.4.5
cryptography==45.0.5
decorator==5.2.1
diff-match-patch==20241021
dill==0.4.0
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-anymail==13.0
django-autoslug==1.9.9
django-celery-beat==2.8.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-import-export==4.3.8
django-model-utils==5.0.0
django-rest-framework==0.1.0
django-rest-knox==5.0.2
django-split-settings==1.3.2
django-storages==1.14.6
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
executing==2.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
moviepy==2.2.1
msgpack==1.1.1
mypy_extensions==1.1.0
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
proglog==0.1.12
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pylint==3.3.7
pytest==8.4.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.3.0
tablib==3.8.0
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
